(CBE) Admin structure refactor.
Rails
=====

* Add a redirect when user doesn't purchased a cbe product;
* Call flash partial in products view;
* Remove unused code from cbes controllers;

VueJs
======

* Move all admin cbe components to cbe/admin folder:
  ** App 		 -> cbe/admin/Home
  ** Show 		 -> cbe/admin/Edit
  ** CBEDetails          -> cbe/admin/Details
  ** CBEIntroductionPage -> cbe/admin/IntroductionPage
  ** CBEQuestion 	 -> cbe/admin/Question
  ** CBEResource 	 -> cbe/admin/Resource
  ** CBESResources 	 -> cbe/admin/Resources
  ** CBEScenario 	 -> cbe/admin/Scenario
  ** CBESction 	 	 -> cbe/admin/Section
