+++ app/javascript/components/cbe/admin/Question.vue
@@ -86,7 +86,7 @@
 
 <script>
 import axios from "axios";
-import TinyEditor from "./TinyEditor";
+import TinyEditor from "../../TinyEditor";
 import { validationMixin } from "vuelidate";
 import { required, numeric, between } from "vuelidate/lib/validators";
