+++ app/javascript/components/cbe/admin/Resource.vue
