+++ app/javascript/components/cbe/admin/Scenario.vue
@@ -40,7 +40,7 @@
 
 <script>
 import axios from "axios";
-import TinyEditor from "./TinyEditor";
+import TinyEditor from "../../TinyEditor";
 import { validationMixin } from "vuelidate";
 import { required } from "vuelidate/lib/validators";
