--- app/javascript/packs/cbes/new.js
@@ -1,7 +1,0 @@
-import Vue from 'vue/dist/vue.js';
-import App from '../../App'
-
-new Vue({
-  template: "",
-  components: { App }
-});
