+++ app/views/admin/cbes/show.html.erb
@@ -1,21 +1,13 @@
 <div class="container">
-
   <section class="pb-7">
     <div class="box-item form-box">
       <div class="row">
         <div class="col-sm-12">
           <div class="cbe-container">
             <div id="cbes-show-view"></div>
-   
           </div>
         </div>
       </div>
     </div>
-
-  </section>
-
-</div>
-
-
   </section>
 </div>
