+++ app/views/footer_pages/media_library.html.haml
@@ -7,7 +7,7 @@
         %p.p-hero.px-lg-6
           Invest in your ACCA exam success today with our question and solution correction packs and mock exams.
 
-
+      =render partial: 'layouts/flash_messages'
 
       %nav.products-nav.mt-6
         %ul.nav.nav-tabs{id: 'productsTabs', role: 'tablist'}
