--- db/seed_data/cbe_question_type.rb
@@ -1,13 +1,0 @@
-CbeQuestionType.where(name: 'Multiple Choice',).first_or_create!(name: 'Multiple Choice', order: 1)
-CbeQuestionType.where(name: 'Multiple Response',).first_or_create!(name: 'Multiple Response', order: 2)
-CbeQuestionType.where(name: 'Complete',).first_or_create!(name: 'Complete', order: 3)
-CbeQuestionType.where(name: 'Fill In The Blank',).first_or_create!(name: 'Fill In The Blank', order: 4)
-CbeQuestionType.where(name: 'Drag & Drop',).first_or_create!(name: 'Complete', order: 5)
-CbeQuestionType.where(name: 'Dropdown List',).first_or_create!(name: 'Dropdown List', order: 6)
-CbeQuestionType.where(name: 'Hot Spot',).first_or_create!(name: 'Complete', order: 7)
-CbeQuestionType.where(name: 'Spreadsheet',).first_or_create!(name: 'Spreadsheet', order: 7)
-CbeQuestionType.where(name: 'Open Text',).first_or_create!(name: 'Open Text', order: 7)
-
-
-
-
