--- db/seed_data/cbe_section_type.rb
@@ -1,3 +1,0 @@
-CbeSectionType.where(name: 'Objective Question').first_or_create!(name: 'Objective Question'); print '.'
-CbeSectionType.where(name: 'Constructed Response').first_or_create!(name: 'Constructed Response'); print '.'
-CbeSectionType.where(name: 'Objective Test Case').first_or_create!(name: 'Objective Test Case'); print '.'
